refactor: 🎨 move mfsu.ts -> mfsu/mfsu.ts
