+++ packages/mfsu/src/index.ts
@@ -2,4 +2,4 @@ export * from './constants';
 export * from './esbuildHandlers/autoCssModules';
 // for independent use `esbuild-loader`
 export { esbuildLoader } from './loader/esbuild';
-export * from './mfsu';
+export * from './mfsu/mfsu';
